+++ src/ncc/ThirdParty/Symfony/process/InputStream.php
@@ -9,9 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace ncc\Symfony\Component\Process;
+namespace ncc\ThirdParty\Symfony\process;
 
-use ncc\Symfony\Component\Process\Exception\RuntimeException;
+use ncc\ThirdParty\Symfony\process\Exception\RuntimeException;
 
 /**
  * Provides a way to continuously write to the input of a Process until the InputStream is closed.
@@ -41,7 +41,7 @@ class InputStream implements \IteratorAggregate
      * @param resource|string|int|float|bool|\Traversable|null $input The input to append as scalar,
      *                                                                stream resource or \Traversable
      */
-    public function write($input)
+    public function write(mixed $input)
     {
         if (null === $input) {
             return;
@@ -68,11 +68,7 @@ class InputStream implements \IteratorAggregate
         return !$this->open;
     }
 
-    /**
-     * @return \Traversable<int, string>
-     */
-    #[\ReturnTypeWillChange]
-    public function getIterator()
+    public function getIterator(): \Traversable
     {
         $this->open = true;